refactor: #12 UserController 변경사항 적용 및 TEST 수정
